+++ fluxo-kmp-conf/src/main/kotlin/PropsAndEnv.kt
@@ -55,8 +55,19 @@ public fun Project.isMaxDebugEnabled(): Provider<Boolean> = envOrPropFlag("MAX_D
 
 internal fun Project.isFluxoVerbose(): Provider<Boolean> = envOrPropFlag("FLUXO_VERBOSE")
 
-/** @TODO: Disables whole artifact processing, not just shrinking; should be renamed. */
-public fun Project.isR8Disabled(): Provider<Boolean> = envOrPropFlag("DISABLE_R8")
+public fun Project.isShrinkerDisabled(): Provider<Boolean> = provider {
+    val disabled = envOrPropFlagValue("DISABLE_R8") ||
+        envOrPropFlagValue("DISABLE_SHRINKER") ||
+        envOrPropFlagValue("DISABLE_PROGUARD")
+
+    val enabled = envOrPropFlagValue("OPTIMIZE") ||
+        envOrPropFlagValue("SHRINK") ||
+        envOrPropFlagValue("ENABLE_R8") ||
+        envOrPropFlagValue("ENABLE_PROGUARD") ||
+        envOrPropFlagValue("ENABLE_SHRINKER")
+
+    disabled && !enabled
+}.memoizeSafe()
 
 public fun Project?.buildNumber(): String? = envOrPropValueLenient("BUILD_NUMBER")